fix: bargain hunter link
Pages with a trailing slash get redirected to a
404 page, that only resolves correctly if
JavaScript is enabled.

This commit removes the trailing slash from the
bargain hunter link, so the page resolves
correctly even if JavaScript is disabled, and
without a brief 404 page even if JavaScript is
enabled.

Related to https://github.com/Merrit/unit_bargain_hunter/issues/84